+++ src/Core/src/Exception/Resolver/UnknownParameterException.php
@@ -4,22 +4,13 @@ declare(strict_types=1);
 
 namespace Spiral\Core\Exception\Resolver;
 
-use Spiral\Core\Exception\Traits\ClosureRendererTrait;
-
 final class UnknownParameterException extends ValidationException
 {
-    use ClosureRendererTrait;
-
-    public function __construct(
+    protected function getValidationMessage(
         \ReflectionFunctionAbstract $reflection,
-        private readonly string $parameter,
-    ) {
+        string $parameter,
+    ): string {
         $pattern = "Unknown named parameter `{$parameter}` `%s` %s.";
-        parent::__construct($this->renderFunctionAndParameter($reflection, $pattern));
-    }
-
-    public function getParameter(): string
-    {
-        return $this->parameter;
+        return $this->renderFunctionAndParameter($reflection, $pattern);
     }
 }